// components/sto-datatable-advanced/types/column-header.type.ts
import { PipeTransform } from '@angular/core';
import { ValueGetter } from '../../../vendor/ngx-datatable/utils/column-prop-getters';
import { TableColumnProp } from '../../../vendor/ngx-datatable/types';


/**
 * Column Type
 * @type {object}
 */
export interface ColumnGroup {

  /**
   * Internal unique id
   * 
   * @type {string}
   * @memberOf TableColumn
   */
  $$id?: string;

  /**
   * Internal for column width distributions
   * 
   * @type {number}
   * @memberOf TableColumn
   */
  $$oldWidth?: number;


  properties?: TableColumnProp[];
  columns?: string[];

  /**
   * The grow factor relative to other columns. Same as the flex-grow
   * API from http =//www.w3.org/TR/css3-flexbox/. Basically; 
   * take any available extra width and distribute it proportionally
   * according to all columns' flexGrow values.
   * 
   * @type {number}
   * @memberOf TableColumn
   */
  flexGrow?: number;

  /**
   * The default width of the column, in pixels
   * 
   * @type {number}
   * @memberOf TableColumn
   */
  width?: number;

  /**
   * Can the column be resized
   * 
   * @type {boolean}
   * @memberOf TableColumn
   */
  resizeable?: boolean;


  /**
   * Can the column be re-arranged by dragging
   * 
   * @type {boolean}
   * @memberOf TableColumn
   */
  draggable?: boolean;

  /**
   * Whether the column can automatically resize to fill space in the table.
   * 
   * @type {boolean}
   * @memberOf TableColumn
   */
  canAutoResize?: boolean;

  /**
   * Column name or label
   * 
   * @type {string}
   * @memberOf TableColumn
   */
  name?: string;


  /**
   * Header template ref
   * 
   * @type {*}
   * @memberOf TableColumn
   */
  headerTemplate?: any;


  /**
   * CSS classes for the header
   * 
   * 
   * @memberOf TableColumn
   */
  headerClass?: string | ((data: any) => string|any);



}